[refactor] Move operatorSourceRow to a new file
- Move operatorSourceRow type to a new file 'row.go' to make
 datastore.go less cluttered.
- operatorSourceRow will be extended later to make it
  concurrent. The refactor will make it easy to browse the code.
